feat(docker): update docker-compose and Intellij config
This update brings several improvements to the docker-compose setup and associated Intellij configuration. The MariaDB image is now set to `latest` and a `healthcheck` is removed as it's no longer needed with the `init: true` directive. The MariaDB container name is simplified to `mariadb`.

The `opencms` service now uses a specific image version `19.0` and includes build configurations for the servlet container and base image. The `depends_on` condition for MariaDB is simplified, and the `links` directive is added for better service discovery.

Volume mounts have been adjusted for both MariaDB and OpenCMS to use a more general path `~/dockermount/opencms-docker-mysql` and `~/dockermount/opencms-docker-webapps` respectively. The `command` for OpenCMS now includes a wait script to ensure the database is available before starting. Environment variables are also updated, including setting `JSONAPI=true`.

Additionally, a new Intellij configuration file `.idea/opencms.xml` has been added to support these changes.